Heroku config

+++ db/migrate/20200919233033_create_orders.rb
@@ -6,7 +6,7 @@ class CreateOrders < ActiveRecord::Migration[6.0]
       t.integer :credit_card_number
       t.date :expiration_date
       t.integer :total_seats_selected
-      t.decimal :total_cost
+      t.integer :total_cost
 
       t.timestamps
     end
